Core: Reuse commit retry failure translation
