+++ core/src/main/java/org/apache/iceberg/BaseTransaction.java
@@ -787,36 +787,11 @@ public class BaseTransaction implements Transaction {
 
   private static CommitFailedException toCommitFailedException(
       RetryExhaustedException ex, Map<String, String> properties) {
-    if (shouldPreserveCommitFailure(ex.getCause())) {
-      return (CommitFailedException) ex.getCause();
-    }
-
     int numRetries =
         PropertyUtil.propertyAsInt(properties, COMMIT_NUM_RETRIES, COMMIT_NUM_RETRIES_DEFAULT);
     int totalTimeoutMs =
         PropertyUtil.propertyAsInt(
             properties, COMMIT_TOTAL_RETRY_TIME_MS, COMMIT_TOTAL_RETRY_TIME_MS_DEFAULT);
-    if (ex.reason() == RetryExhaustedException.Reason.TIMEOUT_EXCEEDED) {
-      return new CommitFailedException(
-          ex,
-          "Commit failed and retry timeout (%d ms) reached. Consider increasing '%s'",
-          totalTimeoutMs,
-          COMMIT_TOTAL_RETRY_TIME_MS);
-    } else {
-      return new CommitFailedException(
-          ex,
-          "Commit failed and retry limit (%d) reached. Consider increasing '%s'",
-          numRetries,
-          COMMIT_NUM_RETRIES);
-    }
-  }
-
-  private static boolean shouldPreserveCommitFailure(Throwable cause) {
-    if (!(cause instanceof CommitFailedException)) {
-      return false;
-    }
-
-    String message = cause.getMessage();
-    return message == null || !message.startsWith("Commit failed: table was updated");
+    return CommitRetry.toCommitFailedException(ex, numRetries, totalTimeoutMs);
   }
 }
