+++ core/src/main/java/org/apache/iceberg/SnapshotProducer.java
@@ -928,35 +928,10 @@ abstract class SnapshotProducer<ThisT> implements SnapshotUpdate<ThisT> {
   }
 
   private CommitFailedException toCommitFailedException(RetryExhaustedException ex) {
-    if (shouldPreserveCommitFailure(ex.getCause())) {
-      return (CommitFailedException) ex.getCause();
-    }
-
     int numRetries = base.propertyAsInt(COMMIT_NUM_RETRIES, COMMIT_NUM_RETRIES_DEFAULT);
     int totalTimeoutMs =
         base.propertyAsInt(COMMIT_TOTAL_RETRY_TIME_MS, COMMIT_TOTAL_RETRY_TIME_MS_DEFAULT);
-    if (ex.reason() == RetryExhaustedException.Reason.TIMEOUT_EXCEEDED) {
-      return new CommitFailedException(
-          ex,
-          "Commit failed and retry timeout (%d ms) reached. Consider increasing '%s'",
-          totalTimeoutMs,
-          COMMIT_TOTAL_RETRY_TIME_MS);
-    } else {
-      return new CommitFailedException(
-          ex,
-          "Commit failed and retry limit (%d) reached. Consider increasing '%s'",
-          numRetries,
-          COMMIT_NUM_RETRIES);
-    }
-  }
-
-  private static boolean shouldPreserveCommitFailure(Throwable cause) {
-    if (!(cause instanceof CommitFailedException)) {
-      return false;
-    }
-
-    String message = cause.getMessage();
-    return message == null || !message.startsWith("Commit failed: table was updated");
+    return CommitRetry.toCommitFailedException(ex, numRetries, totalTimeoutMs);
   }
 
   private static void updateTotal(
