+++ core/src/main/java/org/apache/iceberg/view/SetViewLocation.java
@@ -73,19 +73,7 @@ class SetViewLocation implements UpdateLocation {
               taskOps ->
                   taskOps.commit(base, ViewMetadata.buildFrom(base).setLocation(apply()).build()));
     } catch (RetryExhaustedException e) {
-      if (e.reason() == RetryExhaustedException.Reason.TIMEOUT_EXCEEDED) {
-        throw new CommitFailedException(
-            e,
-            "Commit failed and retry timeout (%d ms) reached. Consider increasing '%s'",
-            totalTimeoutMs,
-            COMMIT_TOTAL_RETRY_TIME_MS);
-      } else {
-        throw new CommitFailedException(
-            e,
-            "Commit failed and retry limit (%d) reached. Consider increasing '%s'",
-            numRetries,
-            COMMIT_NUM_RETRIES);
-      }
+      throw ViewCommitRetry.toCommitFailedException(e, numRetries, totalTimeoutMs);
     }
   }
